requirements.txt
praw
google-genai
python-dotenv 